Moved `Bytes` to `ethane-types`.
