+++ ethane/Cargo.toml
@@ -13,7 +13,6 @@ readme = "../README.md"
 [dependencies]
 ethane-abi = { path = "../ethane-abi", version = "1.0.1" }
 ethane-types = { path = "../ethane-types" }
-hex = "0.4"
 reqwest = { version = "0.11.3", features = ["blocking"] }
 serde = {version = "1", features = ["derive"]}
 serde_json = "1"
